Add an NTManager class
This takes on some of the responsiblity of managing the global
NetworkTables state that was previously just dropped into
NTPublishOperation.  It also sets a logger, which NTPublishOperation did
not do before, that should reduce console spam and let us easily grep
through NetworkTables logs.
