+++ core/src/main/java/edu/wpi/grip/core/operations/Operations.java
@@ -1,7 +1,6 @@
 package edu.wpi.grip.core.operations;
 
 import com.google.common.eventbus.EventBus;
-import edu.wpi.grip.core.Operation;
 import edu.wpi.grip.core.events.OperationAddedEvent;
 import edu.wpi.grip.core.operations.composite.*;
 import edu.wpi.grip.core.operations.networktables.NTPublishOperation;
@@ -32,9 +31,6 @@ public final class Operations {
         eventBus.post(new OperationAddedEvent(new NewPointOperation()));
         eventBus.post(new OperationAddedEvent(new NewSizeOperation()));
         eventBus.post(new OperationAddedEvent(new MatFieldAccessor()));
-
-        Operation publishContours = new NTPublishOperation<>(ContoursReport.class);
-        eventBus.register(publishContours);
-        eventBus.post(new OperationAddedEvent(publishContours));
+        eventBus.post(new OperationAddedEvent(new NTPublishOperation<>(ContoursReport.class)));
     }
 }
